Deleting MasterDetailPage structure for dashboard and replacing it with a simple ContentPage.
