+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay.csproj
@@ -28,15 +28,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="Views\DashboardDetailPage.xaml.cs">
-      <DependentUpon>DashboardDetailPage.xaml</DependentUpon>
-    </Compile>
-    <Compile Update="Views\DashboardMasterPage.xaml.cs">
-      <DependentUpon>DashboardMasterPage.xaml</DependentUpon>
-    </Compile>
-    <Compile Update="Views\DashboardPage.xaml.cs">
-      <DependentUpon>DashboardPage.xaml</DependentUpon>
-    </Compile>
     <Compile Update="Views\Records\AddEditRecordPage.xaml.cs">
       <DependentUpon>AddEditRecordPage.xaml</DependentUpon>
     </Compile>
@@ -58,19 +49,13 @@
   </ItemGroup>
 
   <ItemGroup>
+    <EmbeddedResource Update="Views\DashboardPage.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+    </EmbeddedResource>
     <EmbeddedResource Update="Views\Nutrition\AddEditFoodPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
     <EmbeddedResource Update="Views\Nutrition\AddEditMealPage.xaml">
-      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Views\DashboardPage.xaml">
-      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Views\DashboardDetailPage.xaml">
-      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
-    </EmbeddedResource>
-    <EmbeddedResource Update="Views\DashboardMasterPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
     <EmbeddedResource Update="Views\Nutrition\FoodsPage.xaml">
