--- NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Views/DashboardMasterPage.xaml
@@ -1,50 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d"
-             x:Class="NeverSkipLegDay.Views.DashboardMasterPage"
-             Title="Master">
-  <StackLayout BackgroundColor="{StaticResource CardBackgroundColor}">
-    <ListView x:Name="MenuItemsListView"
-      SeparatorVisibility="Default"
-      HasUnevenRows="true"
-      ItemsSource="{Binding MenuItems}"
-      SeparatorColor="{StaticResource SeparatorColor}"
-      ItemSelected="OnItemSelected">
-      <ListView.Header>
-        <StackLayout BackgroundColor="{StaticResource CardBackgroundColor}">
-          <Frame CornerRadius="100"
-            HeightRequest="200"
-            WidthRequest="200"
-            HorizontalOptions="Center"
-            Padding="0"
-            IsClippedToBounds="True"
-            Margin="20">
-            <Image Source="https://i.imgur.com/PoQiWKf.jpg"
-              HorizontalOptions="Center"
-              VerticalOptions="Center" />
-          </Frame>
-        </StackLayout>
-      </ListView.Header>
-      <ListView.ItemTemplate>
-        <DataTemplate>
-          <ViewCell>
-            <StackLayout
-              BackgroundColor="{StaticResource CardBackgroundColor}"
-              Padding="15,10" 
-              HorizontalOptions="FillAndExpand" 
-              VerticalOptions="FillAndExpand">
-              <Label VerticalOptions="FillAndExpand" 
-                VerticalTextAlignment="Center" 
-                Text="{Binding Title}"
-                FontSize="20"
-                TextColor="{StaticResource CardTextColor}"/>
-            </StackLayout>
-          </ViewCell>
-        </DataTemplate>
-      </ListView.ItemTemplate>
-    </ListView>
-  </StackLayout>
-</ContentPage>
